+++ addOns/pscanrules/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Cache-control scan rule no longer checks if Pragma is set or not.
 - Maintenance changes.
+- The Timestamp Disclosure scan rule now excludes values in "Report-To" or "NEL" headers (Issue 6493).
 
 ## [33] - 2021-01-29
 ### Added